+++ visualization/app/codeCharta/state/store/appStatus/hoveredBuildingPath/hoveredBuildingPath.selector.ts
@@ -1,3 +1,0 @@
-import { CcState } from "../../store"
-
-export const hoveredBuildingPathSelector = (ccState: CcState) => ccState.appStatus.hoveredBuildingPath