+++ visualization/app/codeCharta/state/store/appStatus/hoveredNodeId/hoveredNodeId.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { appStatusSelector } from "../appStatus.selector"
+
+export const hoveredNodeIdSelector = createSelector([appStatusSelector], appStatus => appStatus.hoveredNodeId)