+++ visualization/app/codeCharta/ui/dialog/dialog.service.ts
@@ -1,7 +1,6 @@
 import { dialogDownloadComponent } from "./dialog.download.component"
 import { dialogGlobalSettingsComponent } from "./dialog.globalSettings.component"
 import { addScenarioSettingsComponent } from "./dialog.addScenarioSettings.component"
-import { addCustomConfigSettingsComponent } from "./dialog.addCustomConfigSettings.component"
 import { CCFileValidationResult } from "../../util/fileValidator"
 import { dialogChangelogComponent } from "./dialog.changelog.component"
 
@@ -20,10 +19,6 @@ export class DialogService {
 
 	showAddScenarioSettings() {
 		this.showCustomDialog(addScenarioSettingsComponent)
-	}
-
-	showAddCustomConfigSettings() {
-		this.showCustomDialog(addCustomConfigSettingsComponent)
 	}
 
 	showChangelogDialog() {