+++ visualization/app/codeCharta/ui/ui.ts
@@ -12,7 +12,6 @@ import "./edgeSettingsPanel/edgeSettingsPanel.module"
 import "./edgeChooser/edgeChooser.module"
 import "./presentationModeButton/presentationModeButton.module"
 import "./centerMapButton/centerMapButton.module"
-import "./searchPanel/searchPanel.module"
 import "./areaSettingsPanel/areaSettingsPanel.module"
 import "./codeMap/codeMap.module"
 import "./colorSettingsPanel/colorSettingsPanel.module"
@@ -21,7 +20,6 @@ import "./fileChooser/fileChooser.module"
 import "./fileExtensionBar/fileExtensionBar.module"
 import "./heightSettingsPanel/heightSettingsPanel.module"
 import "./metricChooser/metricChooser.module"
-import "./metricType/metricType.module"
 import "./rangeSlider/rangeSlider.module"
 import "./resetSettingsButton/resetSettingsButton.module"
 import "./filePanel/filePanel.module"
@@ -32,12 +30,6 @@ import "./viewCube/viewCube.module"
 import "./layoutSelection/layoutSelection.module"
 import "./maxTreeMapFiles/maxTreeMapFiles.module"
 import "./sharpnessModeSelector/sharpnessModeSelector.module"
-import { MetricDeltaSelectedComponent } from "./attributeSideBar/metricDeltaSelected/metricDeltaSelected.component"
-import { MapTreeViewComponent } from "./mapTreeView/mapTreeView.component"
-import { SortingOptionComponent } from "./sortingOption/sortingOption.component"
-import { MatchingFilesCounterComponent } from "./matchingFilesCounter/matchingFilesCounter.component"
-import { SortingButtonComponent } from "./sortingButton/sortingButton.component"
-import { AttributeTypeSelectorComponent } from "./attributeSideBar/attributeTypeSelector/attributeTypeSelector.component"
 import { Export3DMapButtonComponent } from "./export3DMapButton/export3DMapButton.component"
 import { LegendPanelComponent } from "./legendPanel/legendPanel.component"
 
@@ -54,8 +46,6 @@ angular
 		"app.codeCharta.ui.edgeChooser",
 		"app.codeCharta.ui.presentationModeButton",
 		"app.codeCharta.ui.centerMapButton",
-		"app.codeCharta.ui.searchPanel",
-		"app.codeCharta.ui.metricType",
 		"app.codeCharta.ui.areaSettingsPanel",
 		"app.codeCharta.ui.codeMap",
 		"app.codeCharta.ui.colorSettingsPanel",
@@ -64,7 +54,6 @@ angular
 		"app.codeCharta.ui.fileExtensionBar",
 		"app.codeCharta.ui.heightSettingsPanel",
 		"app.codeCharta.ui.metricChooser",
-		"app.codeCharta.ui.metricType",
 		"app.codeCharta.ui.rangeSlider",
 		"app.codeCharta.ui.resetSettingsButton",
 		"app.codeCharta.ui.ribbonBar",
@@ -76,11 +65,5 @@ angular
 		"app.codeCharta.ui.maxTreeMapFiles",
 		"app.codeCharta.ui.filePanel"
 	])
-	.directive("ccSortingButton", downgradeComponent({ component: SortingButtonComponent }))
-	.directive("ccSortingOption", downgradeComponent({ component: SortingOptionComponent }))
-	.directive("ccMetricDeltaSelected", downgradeComponent({ component: MetricDeltaSelectedComponent }))
-	.directive("ccMapTreeView", downgradeComponent({ component: MapTreeViewComponent }))
-	.directive("ccMatchingFilesCounter", downgradeComponent({ component: MatchingFilesCounterComponent }))
-	.directive("ccAttributeTypeSelector", downgradeComponent({ component: AttributeTypeSelectorComponent }))
 	.directive("ccExportThreedMapButton", downgradeComponent({ component: Export3DMapButtonComponent }))
 	.directive("ccLegendPanel", downgradeComponent({ component: LegendPanelComponent }))